requirements.txt
# --- Core framework ---
fastapi>=0.115.0
uvicorn>=0.30.0

# --- Environment & utilities ---
python-dotenv>=1.0.1
requests>=2.32.0

# --- GitHub REST integration ---
PyGithub>=2.3.0

# --- Google Gemini / Generative AI ---
google-generativeai>=0.8.5

# --- Optional: faster JSON, event loop, reload tools ---
ujson>=5.10.0
uvloop>=0.19.0
watchfiles>=0.20.0

# --- Type system (FastAPI + Pydantic v2 deps) ---
pydantic>=2.7.0
pydantic-core>=2.20.0

# --- Optional debugging/logging ---
rich>=13.7.0
tqdm>=4.65.0

